Add point presenter
